Added a fetchRankedAddressBookWithCompletionHandler method
In general, I would expect apps using the SDK to want a callback/completion handler
block when making requests to the SDK, allowing them to perform any necessary tasks
once the data is ready.

This is a pattern used by many SDKs that need to make requests to the network
to get/post data.  I would recommend following this pattern.

If this pattern is not desireable, other options include:
  * creating a delegate protocol for handling data
  * using KVO
  * using notifications